refactor: rename {TextDocument => }FeatureProviderRegistry
